feat: add support for flat configs
