+++ src/__tests__/rules.test.ts
@@ -1,6 +1,6 @@
 import { existsSync } from 'fs';
 import { resolve } from 'path';
-import plugin from '../';
+import plugin from '../legacy';
 
 const numberOfRules = 53;
 const ruleNames = Object.keys(plugin.rules);
