+++ src/configs/style.ts
@@ -1,0 +1,12 @@
+import legacy from '../legacy';
+import all from './all';
+
+const [snap, test] = all;
+
+export default [
+  snap,
+  {
+    ...test,
+    rules: legacy.configs.style.rules,
+  },
+];
